+++ src/services/apimap-field-builder.js
@@ -149,6 +149,11 @@ function ApimapFieldBuilder(model, column, options) {
       schema.enums = column.values;
     }
 
+    // NOTICE: Create enums from sub-type (for ['Enum'] type).
+    if (Array.isArray(schema.type) && schema.type[0] === 'Enum') {
+      schema.enums = column.type.type.values;
+    }
+
     if (isRequired()) {
       schema.isRequired = true;
     }